// src/app/models/Tesoreria/mtmInstrumentoHistorico.ts
export class MTMInstrumentoHistorico{
    t467_id: number;
    t467_date: Date;
    cod_ci: string;
    t463_currency_rec: string;
    t467_mtm_rec: number;
    t463_currency_del: string;
    t467_mtm_del: number;
    mtm_rec_conv: number;
    mtm_del_conv: number;
    mtm_total: number;
    t463_type_ci: string;
    t463_id_subsidiary: number;
    t463_id_coverage_type: number;
    t018_code_alpha2: string;
}